+++ test/raxx/router_test.exs
@@ -150,20 +150,22 @@ defmodule Raxx.RouterTest do
     # HEAD MIDDLEA
     defmodule AuthorizationMiddleware do
       use Raxx.Middleware
+      alias Raxx.Server
 
       @impl Raxx.Middleware
-      def process_head(head, :pass, next) do
-        {parts, inner_server} = Server.handle_head(inner_server, request)
-        {parts, :pass, inner_server}
+      def process_head(request, :pass, next) do
+        {parts, next} = Server.handle_head(next, request)
+        {parts, :pass, next}
       end
 
-      def process_head(head, :stop, next) do
-        Raxx.response(:forbidden)
+      def process_head(_request, :stop, next) do
+        {[Raxx.response(:forbidden)], :stop, next}
       end
     end
 
     defmodule TestHeaderMiddleware do
       use Raxx.Middleware
+      alias Raxx.Server
 
       @impl Raxx.Middleware
       def process_head(request, state, inner_server) do
@@ -193,7 +195,13 @@ defmodule Raxx.RouterTest do
         {parts, state, inner_server}
       end
 
-      def add_header(_, _vale) do
+      def add_header([response = %Raxx.Response{} | rest], state) do
+        response = Raxx.set_header(response, "x-test", state)
+        [response | rest]
+      end
+
+      def add_header(parts, _state) when is_list(parts) do
+        parts
       end
     end
 
@@ -201,7 +209,7 @@ defmodule Raxx.RouterTest do
       use Raxx.Router
 
       # Test with HEAD middleware
-      section([{TestHeaderMiddleware, "bonus"}], [
+      section([{TestHeaderMiddleware, "compile-time"}], [
         {%{method: :GET, path: []}, HomePage}
       ])
 
@@ -215,13 +223,100 @@ defmodule Raxx.RouterTest do
       ])
 
       def private(state) do
-        [{AuthorizationMiddleware, state.authorization}]
+        [
+          {TestHeaderMiddleware, "run-time"},
+          {AuthorizationMiddleware, state.authorization}
+        ]
       end
+    end
 
-      # MOVE ALL TESTS here
-      # check auth config
-      # check header
-      # Test with Auth middleware and HEAD
+    test "will route to homepage" do
+      request = Raxx.request(:GET, "/")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert "Home page" == response.body
+    end
+
+    test "will route to fixed segment" do
+      request = Raxx.request(:GET, "/users")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert "Users page" == response.body
+    end
+
+    test "will route to variable segment path" do
+      request = Raxx.request(:GET, "/users/34")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert "User page 34" == response.body
+    end
+
+    test "will route on method" do
+      request = Raxx.request(:POST, "/users")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert "User created " == response.body
+    end
+
+    test "will forward whole request to controller" do
+      request =
+        Raxx.request(:POST, "/users")
+        |> Raxx.set_body(true)
+
+      {[], state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      {[], state} = SectionRouter.handle_data("Bob", state)
+      {[response], _state} = SectionRouter.handle_tail([], state)
+      assert "User created Bob" == response.body
+    end
+
+    test "will route on catch all" do
+      request = Raxx.request(:GET, "/random")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert "Not found" == response.body
+    end
+
+    @tag :skip
+    test "adds the action module to logger metadata" do
+      request = Raxx.request(:GET, "/")
+      _ = SectionRouter.handle_head(request, %{authorization: :pass})
+      metadata = Logger.metadata()
+      assert "Raxx.RouterTest.HomePage" = Keyword.get(metadata, :"raxx.action")
+      assert "%{method: :GET, path: []}" = Keyword.get(metadata, :"raxx.route")
+    end
+
+    test "will raise return error if fails to route simple request" do
+      request = Raxx.request(:GET, "/invalid")
+
+      assert_raise ReturnError, fn ->
+        SectionRouter.handle_head(request, %{authorization: :pass})
+      end
+    end
+
+    test "will raise return error if fails to route streamed request" do
+      request =
+        Raxx.request(:POST, "/invalid")
+        |> Raxx.set_body(true)
+
+      {[], state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      {[], state} = SectionRouter.handle_data("Bob", state)
+
+      assert_raise ReturnError, fn ->
+        SectionRouter.handle_tail([], state)
+      end
+    end
+
+    test "middleware as list is applied" do
+      request = Raxx.request(:GET, "/")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert "compile-time" == Raxx.get_header(response, "x-test")
+    end
+
+    test "middleware as function is applied" do
+      request = Raxx.request(:GET, "/users")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :pass})
+      assert 200 == response.status
+      assert "run-time" == Raxx.get_header(response, "x-test")
+
+      request = Raxx.request(:GET, "/users")
+      {[response], _state} = SectionRouter.handle_head(request, %{authorization: :stop})
+      assert 403 == response.status
+      assert "run-time" == Raxx.get_header(response, "x-test")
     end
   end
 end